+++ src/main/java/com/lucas/cadastrodeempregados/empregados/EmpregadoRepository.java
@@ -2,5 +2,5 @@ package com.lucas.cadastrodeempregados.empregados;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmpregadoRepository extends JpaRepository<Empregado, Integer> {
+public interface EmpregadoRepository extends JpaRepository<Empregado, Long> {
 }
