+++ src/main/java/com/lucas/cadastrodeempregados/tarefas/TarefaRepository.java
@@ -2,5 +2,5 @@ package com.lucas.cadastrodeempregados.tarefas;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TarefaRepository extends JpaRepository<Tarefa, Integer> {
+public interface TarefaRepository extends JpaRepository<Tarefa, Long> {
 }
